Skip attribute if it's in omitted attributes
There may be times when you want to use an existing script that has
attributes that should be ignored on the next dump. For example you are
no longer using a column in a table but haven't been able to drop it yet
replicate will still dump that column until it no longer exists. The
loader may not want to use this column. This is especially useful for
Rails 4.2 where you cannot write virtual attributes or attributes that
no longer exist. This way we can ignore the columns on load the same way
they're ignored on dump.