Re-tool the ActiveRecord aspect of this library to be strict
This change changes from a Dump-Everything default to a Dump-Nothing.
Now, each ActiveRecord model that should be dumped needs to be flagged
as such `replicate_enable`, each of that model's attributes, outside of
the primary key, needs to be explicitly mentioned, and every association
that should be included in a dump needs to be explicitly mentioned.